test: Add comprehensive unit tests for microservices and improve error handling
Add complete unit test coverage for Patient, Assessment, and Note microservices including both controller and service layers with edge cases, boundary values, and error scenarios.

Improve exception handling with enhanced GlobalExceptionHandler, better 400/404/500 error responses, and fixed custom FeignException handling across services.

Update Maven pom.xml configurations to support test execution, add JaCoCo plugin for code coverage, and enhance Surefire plugin setup for better test reporting.

Minor code improvements include removing unused functions, cleaning up imports, improving code consistency, and enhancing logging messages.

Add GitHub Actions workflow for automated testing with parallel execution and independent test jobs to prevent cascade failures.

+++ assessment/src/main/java/com/medilabo/solutions/assessment/client/PatientServiceClient.java
@@ -6,11 +6,10 @@ import org.springframework.web.bind.annotation.PathVariable;
 
 import com.medilabo.solutions.assessment.dto.PatientDto;
 
-//TODO on passe par gateway ici aussi ou pas ?
 @FeignClient(name = "patient")
 public interface PatientServiceClient {
 
     @GetMapping("/api/patient/{id}")
-    PatientDto getPatientById(@PathVariable("id") Long patientId);
+    PatientDto getPatientById(@PathVariable("id") int patientId);
 
 }
